pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.40",
    "streamlit>=1.45.0",
]
